Seperate code between Core and View directory
