+++ TarponSimulator2017.userprefs
@@ -1,11 +1,10 @@
 ﻿<Properties StartupConfiguration="{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}|Default" StartupItem="TarponSimulator2017/TarponSimulator2017.csproj">
-  <MonoDevelop.Ide.Workbench ActiveDocument="TarponSimulator2017/TarponGame.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="TarponSimulator2017/View/Player.cs">
     <Files>
-      <File FileName="TarponSimulator2017/MonoGame.Framework.dll.config" Line="1" Column="1" />
-      <File FileName="TarponSimulator2017/TarponGame.cs" Line="1" Column="1" />
-      <File FileName="TarponSimulator2017/Sprite.cs" Line="1" Column="1" />
-      <File FileName="TarponSimulator2017/Player.cs" Line="1" Column="1" />
-      <File FileName="TarponSimulator2017/packages.config" Line="1" Column="1" />
+      <File FileName="TarponSimulator2017/TarponGame.cs" Line="83" Column="1" />
+      <File FileName="TarponSimulator2017/Core/Boat.cs" Line="35" Column="16" />
+      <File FileName="TarponSimulator2017/View/Sprite.cs" Line="16" Column="3" />
+      <File FileName="TarponSimulator2017/View/Player.cs" Line="41" Column="3" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|x86" />
