+++ TarponSimulator2017/TarponSimulator2017.sln
@@ -1,0 +1,17 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TarponSimulator2017", "TarponSimulator2017.csproj", "{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x86 = Debug|x86
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}.Debug|x86.ActiveCfg = Debug|x86
+		{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}.Debug|x86.Build.0 = Debug|x86
+		{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}.Release|x86.ActiveCfg = Release|x86
+		{4DC88FC1-C0C7-4542-BCB4-66EB064779CD}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+EndGlobal
